refactor: move `is_initial_block_download` to snapshot
