+++ sync/Cargo.toml
@@ -19,20 +19,21 @@ ckb-network = { path = "../network", version = "= 0.41.0-pre" }
 ckb-logger = {path = "../util/logger", version = "= 0.41.0-pre"}
 ckb-metrics = {path = "../util/metrics", version = "= 0.41.0-pre"}
 ckb-util = { path = "../util", version = "= 0.41.0-pre" }
-faketime = "0.2.0"
-bitflags = "1.0"
 ckb-verification = { path = "../verification", version = "= 0.41.0-pre" }
 ckb-verification-traits = { path = "../verification/traits", version = "= 0.41.0-pre" }
 ckb-chain-spec = { path = "../spec", version = "= 0.41.0-pre" }
 ckb-channel = { path = "../util/channel", version = "= 0.41.0-pre" }
 ckb-traits = { path = "../traits", version = "= 0.41.0-pre" }
-lru = "0.6.0"
-sentry = { package = "ckb-sentry", version = "0.21.0", optional = true }
-futures = "0.3"
 ckb-error = {path = "../error", version = "= 0.41.0-pre"}
 ckb-tx-pool = { path = "../tx-pool", version = "= 0.41.0-pre" }
+sentry = { package = "ckb-sentry", version = "0.21.0", optional = true }
+ckb-constant = { path = "../util/constant", version = "= 0.41.0-pre" }
+lru = "0.6.0"
+futures = "0.3"
 governor = "0.3.1"
 tempfile = "3.0"
+faketime = "0.2.0"
+bitflags = "1.0"
 
 [dev-dependencies]
 ckb-test-chain-utils = { path = "../util/test-chain-utils", version = "= 0.41.0-pre" }
