+++ sync/src/synchronizer/headers_process.rs
@@ -1,7 +1,8 @@
 use crate::block_status::BlockStatus;
 use crate::synchronizer::Synchronizer;
 use crate::types::{ActiveChain, SyncShared};
-use crate::{Status, StatusCode, MAX_HEADERS_LEN};
+use crate::{Status, StatusCode};
+use ckb_constant::sync::MAX_HEADERS_LEN;
 use ckb_error::Error;
 use ckb_logger::{debug, log_enabled, warn, Level};
 use ckb_network::{CKBProtocolContext, PeerIndex};
