+++ test/Cargo.toml
@@ -21,6 +21,7 @@ ckb-hash = { path = "../util/hash", version = "= 0.41.0-pre" }
 ckb-util = { path = "../util", version = "= 0.41.0-pre" }
 ckb-chain-spec = { path = "../spec", version = "= 0.41.0-pre" }
 ckb-crypto = { path = "../util/crypto", version = "= 0.41.0-pre" }
+ckb-dao = { path = "../util/dao", version = "= 0.41.0-pre" }
 ckb-dao-utils = { path = "../util/dao/utils", version = "= 0.41.0-pre" }
 ckb-test-chain-utils = { path = "../util/test-chain-utils", version = "= 0.41.0-pre" }
 ckb-resource = { path = "../resource", version = "= 0.41.0-pre" }
@@ -30,6 +31,7 @@ ckb-logger = { path = "../util/logger", version = "= 0.41.0-pre" }
 ckb-logger-config = { path = "../util/logger-config", version = "= 0.41.0-pre" }
 ckb-logger-service = { path = "../util/logger-service", version = "= 0.41.0-pre" }
 ckb-error = { path = "../error", version = "= 0.41.0-pre" }
+ckb-constant = { path = "../util/constant", version = "= 0.41.0-pre" }
 tempfile = "3.0"
 reqwest = { version = "0.10.9", features = ["blocking", "json"] }
 rand = "0.7"
