+++ util/constant/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ckb-constant"
+version = "0.41.0-pre"
+license = "MIT"
+authors = ["Nervos Core Dev <dev@nervos.org>"]
+edition = "2018"
+description = "Ckb constant container"
+homepage = "https://github.com/nervosnetwork/ckb"
+repository = "https://github.com/nervosnetwork/ckb"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
