+++ util/snapshot/Cargo.toml
@@ -22,3 +22,5 @@ arc-swap = "0.4"
 ckb-reward-calculator = { path = "../reward-calculator", version = "= 0.41.0-pre" }
 ckb-db-schema = { path = "../../db-schema", version = "= 0.41.0-pre" }
 ckb-freezer = { path = "../../freezer", version = "= 0.41.0-pre" }
+ckb-constant = { path = "../constant", version = "= 0.41.0-pre" }
+faketime = "0.2.0"
